reboot機能の修正

+++ src/pal-admin/shells/reboot-check.sh
@@ -2,5 +2,10 @@
 DIR=$(dirname "$0")
 if [ -f "$DIR/reboot.flag" ]; then
     rm "$DIR/reboot.flag"
-    sudo reboot
+    # ファイルが正常に削除されたか確認
+    if [ ! -f "$DIR/reboot.flag" ]; then
+        sudo reboot
+    else
+        echo "Failed to delete reboot.flag file. Aborting reboot."
+    fi
 fi
